Moved popup processing into separate module
